requirements.txt
fastapi
uvicorn
torch
torchvision
Pillow
opencv-python